+++ sdk/textanalytics/azure-ai-textanalytics/azure/ai/textanalytics/_response_handlers.py
@@ -36,6 +36,7 @@ from ._models import (
     RecognizeCustomEntitiesResult,
     ClassifyDocumentResult,
     ActionPointerKind,
+    AbstractSummaryResult,
 )
 
 
@@ -127,6 +128,15 @@ def prepare_result(func):
         return wrapper(*args, ordering_function=order_results)
 
     return choose_wrapper
+
+
+@prepare_result
+def abstract_summary_result(
+    summary, results, *args, **kwargs
+):  # pylint: disable=unused-argument
+    return AbstractSummaryResult._from_generated(  # pylint: disable=protected-access
+        summary
+    )
 
 
 @prepare_result
@@ -312,6 +322,8 @@ def _get_deserialization_callback_from_task_type(task_type):  # pylint: disable=
         return classify_document_result
     if task_type == _AnalyzeActionsType.ANALYZE_HEALTHCARE_ENTITIES:
         return healthcare_result
+    if task_type == _AnalyzeActionsType.ABSTRACT_SUMMARY:
+        return abstract_summary_result
     return key_phrases_result